test(internal/database): move join test to test_impl
